requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Werkzeug==2.3.7
requests==2.31.0
urllib3==2.0.7
SQLAlchemy>=2.0.25
schedule==1.2.0
